Add files via upload
The code is a little messy (sorry) but what I did is that I made some changes to both to the grid and to the bull (just minor changes) and then I added Util.py because python gives me an error saying something about having a library issue so I was force to create a util.py but I don't know if we need it (please help). 

But MDProbot! WOOOO please check it out